requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
async-timeout==5.0.1
asyncssh==2.20.0
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
certifi==2025.1.31
certipy==0.2.1
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
distro==1.9.0
dnspython==2.7.0
feedparser==6.0.11
Flask==3.1.0
fqdn==1.5.1
frozenlist==1.5.0
gevent==24.11.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-ws==0.7.1
idna==3.10
isoduration==20.11.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
Mako==1.3.9
markdownify==0.14.1
MarkupSafe==3.0.2
multidict==6.1.0
oauthlib==3.2.2
openai==1.63.2
packaging==24.2
pamela==1.2.0
prometheus_client==0.21.1
propcache==0.3.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pypandoc==1.15
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.23.1
setuptools==75.8.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
uri-template==1.3.0
urllib3==2.3.0
webcolors==24.11.1
websocket==0.2.1
websockets==15.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
